isolate response merging logic
